strategy skills: rename _mcp.py -> mcp_client.py + declare it (prod-install robustness)
Same fix as the read skills: the vendored MCP helper (imported at runtime by
discover.py and ops' deploy/status/close.py) can be dropped by a selective
installer that strips '_'-prefixed files or fetches only declared files.
- git mv _mcp.py -> mcp_client.py (discover + ops); rewired all imports.
- updated doc refs (ops SKILL.md, references/lifecycle.md).
- added an Install note to both SKILL.md (install the whole scripts/ dir).
Imports resolve; discover's 164 engine/CLI/fleet tests pass (the lone
test_scenarios failure is a pre-existing missing-catalog.json path issue,
unrelated to this change).
