+++ senpi-strategy-discover/scripts/mcp_client.py
