+++ senpi-strategy-ops/scripts/mcp_client.py
